+++ zerver/lib/bulk_create.py
@@ -7,6 +7,7 @@ from zerver.models import Realm, Stream, UserProfile, \
     Subscription, Recipient, RealmAuditLog
 from zerver.lib.create_user import create_user_profile, \
     get_display_email_address
+from zerver.lib.streams import render_stream_description
 
 def bulk_create_users(realm: Realm,
                       users_raw: Set[Tuple[str, str, str, bool]],
@@ -111,7 +112,6 @@ def bulk_create_streams(realm: Realm,
             options['history_public_to_subscribers'] = (
                 not options.get("invite_only", False) and not realm.is_zephyr_mirror_realm)
         if name.lower() not in existing_streams:
-            from zerver.lib.actions import render_stream_description
             streams_to_create.append(
                 Stream(
                     realm=realm,
